split out deckmap component
